test(endpoint-json-feed): use node test runner
